+++ geulnamu-backend/src/main/java/com/geulnamu/controller/auth/LoginController.java
@@ -1,11 +1,11 @@
-package com.geulnamu.controller.member;
+package com.geulnamu.controller.auth;
 
-import com.geulnamu.controller.member.dto.response.LoginResponseDTO;
+import com.geulnamu.controller.auth.dto.response.LoginResponseDTO;
 import com.geulnamu.domain.shared.enums.Level;
 import com.geulnamu.global.response.BaseResponse;
 import com.geulnamu.infrastructure.annotation.AuthToken;
 import com.geulnamu.infrastructure.annotation.AccessLevel;
-import com.geulnamu.service.member.LoginFacade;
+import com.geulnamu.service.auth.LoginFacade;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
