+++ geulnamu-backend/src/main/java/com/geulnamu/controller/auth/dto/response/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package com.geulnamu.controller.member.dto.response;
+package com.geulnamu.controller.auth.dto.response;
 
 public record LoginResponseDTO(String accessToken, boolean newMember) {
     public static LoginResponseDTO of(String accessToken, boolean newMember) {
